서버 레이아웃 작성으로 로딩 파일 간소화
